Complete prompt movie separation
